+++ libs/ledgerjs/packages/hw-app-xrp/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @ledgerhq/hw-app-xrp
 
+## 6.28.6
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.30.6
+
+## 6.28.6-next.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @ledgerhq/hw-transport@6.30.6-next.0
+
 ## 6.28.5
 
 ### Patch Changes